refactor: move media paths under runtime media
